backend/requirements.txt
# ── Web Framework ──────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# ── Database ───────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# ── Auth ───────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# ── Config / Validation ────────────────────────
pydantic==2.7.1
pydantic-settings==2.3.1
python-dotenv==1.0.1

# ── AI / Agents ────────────────────────────────
anthropic==0.28.0
langchain==0.2.5
langchain-anthropic==0.1.13
langchain-community==0.2.5
langgraph==0.1.5

# ── JIRA Integration ───────────────────────────
jira==3.8.0

# ── Git Integration ────────────────────────────
gitpython==3.1.43

# ── Async / Tasks ──────────────────────────────
celery[redis]==5.4.0
redis==5.0.7
websockets==12.0

# ── HTTP Client ────────────────────────────────
httpx==0.27.0

# ── Dev / Testing ──────────────────────────────
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0
ruff==0.4.9
mypy==1.10.0
pre-commit==3.7.1
